+++ helper-scripts/autoupdate.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
-#V0.0.17 - Beta
-
-# skip autoupdate in docker
-if [[ ! -z "$ETV_FILLER_DOCKER" ]]
-then
-    return 0 2>/dev/null || exit 0
-fi
+#V0.0.18 - Beta
 
 # load in configuration variables
 . config-temp.conf
+
+git fetch |& tee $workdir/update
+
+if [[ -s $workdir/update ]];
+then
+
+
 #for test
 #test variable run yes/no
 #convert variable to lowercase
@@ -16,9 +17,10 @@ autoupdate1=$(echo $autoupdate | tr '[:upper:]' '[:lower:]')
 if [[ $autoupdate1 = yes ]]
 then
 echo '#!/bin/bash' >> /tmp/ErsatzTV-Filler-autoupdate.sh
-echo cd $scriptdir
+echo cd $scriptdir >> /tmp/ErsatzTV-Filler-autoupdate.sh
 echo git pull >> /tmp/ErsatzTV-Filler-autoupdate.sh
-echo rm \$0 >> /tmp/ErsatzTV-Filler-autoupdate.sh
+echo "rm -f $scriptdir/running.txt" >> /tmp/ErsatzTV-Filler-autoupdate.sh
+echo exec $scriptdir/generator.sh >> /tmp/ErsatzTV-Filler-autoupdate.sh
 
 chmod +x /tmp/ErsatzTV-Filler-autoupdate.sh
 
@@ -29,7 +31,10 @@ cd $scriptdir
 
 exec /tmp/ErsatzTV-Filler-autoupdate.sh
 else
-rm -f $helperdir/config-temp.conf
-cd $scriptdir
-exit 0
+cd $helperdir
+./weather.sh
+fi
+else
+cd $helperdir
+./weather.sh
 fi