+++ helper-scripts/cleanup.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
-#V0.0.17 - Beta
+#V0.0.18 - Beta
 
 # load in configuration variables
 . config-temp.conf
 
 #General cleanup
 
+echo "Starting clean up"
+
 rm -f $weatherdir/*
 rm -r $workdir/*
 rm -f $scriptdir/running.txt
-
-./autoupdate.sh
+rm -f $workdir/update
+find $log_location -mtime +$log_days -exec rm -f {} \;
+echo "Finished cleaning up"
+exit 0