+++ helper-scripts/news.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#V0.0.17 - Beta
+#V0.0.18 - Beta
 
 # load in configuration variables
 . config-temp.conf
@@ -13,13 +13,19 @@ fi
 processnews1=$(echo $processnews | tr '[:upper:]' '[:lower:]')
 if [[ $processnews1 = yes ]]
 then
-
 #news
 
 #audio
-randomNumber3=$audionumber
-randomNumber4=$audionumber
-randomNumber5=$audionumber
+if [[ $audioamount = 1 ]]
+then
+randomNumber3=1
+randomNumber4=1
+randomNumber5=1
+else
+  randomNumber3=$(shuf -i 1-$audioamount -n 1 --repeat)
+  randomNumber4=$(shuf -i 1-$audioamount -n 1 --repeat)
+  randomNumber5=$(shuf -i 1-$audioamount -n 1 --repeat)
+fi
 audio3=$(head -n $randomNumber3 $workdir/music.txt | tail -n 1)
 audio4=$(head -n $randomNumber4 $workdir/music.txt | tail -n 1)
 audio5=$(head -n $randomNumber5 $workdir/music.txt | tail -n 1)
@@ -112,11 +118,14 @@ cat $workdir/news22.txt | sed 's/\%/\\%/g' >> $workdir/news.txt
 
 
 
-
+# Maths for fade
+newsvideofadeoutstart=$(echo `expr $newsduration1 - $newsvideofadeoutduration` | bc)
+newsaudiofadeoutstart=$(echo `expr $newsduration1 - $newsaudiofadeoutduration` | bc)
 
 # Generate Video
 
-ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio3" -shortest -vf "drawtext=textfile='$workdir/news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $output/news-v1.mp4
+ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio3" -shortest -vf "drawtext=textfile='$workdir/news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $workdir/news-v1.mp4
+ffmpeg -y -i $workdir/news-v1.mp4 -vf "fade=t=in:st=0:d=$newsvideofadeinduration,fade=t=out:st=$newsvideofadeoutstart:d=$newsvideofadeoutduration" -af "afade=t=in:st=0:d=$newsaudiofadeinduration,afade=t=out:st=$newsaudiofadeoutstart:d=$newsaudiofadeoutduration" $output/news-v1.mp4
 touch $output/news-v1.mp4
 
 
@@ -161,7 +170,8 @@ cat $workdir/optional1-news22.txt | sed 's/\%/\\%/g' >> $workdir/optional1-news.
 
 # Generate Video
 
-ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio4" -shortest -vf "drawtext=textfile='$workdir/optional1-news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $output/news-v2.mp4
+ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio4" -shortest -vf "drawtext=textfile='$workdir/optional1-news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $workdir/news-v2.mp4
+ffmpeg -y -i $workdir/news-v2.mp4 -vf "fade=t=in:st=0:d=$newsvideofadeinduration,fade=t=out:st=$newsvideofadeoutstart:d=$newsvideofadeoutduration" -af "afade=t=in:st=0:d=$newsaudiofadeinduration,afade=t=out:st=$newsaudiofadeoutstart:d=$newsaudiofadeoutduration" $output/news-v2.mp4
 touch $output/news-v2.mp4
 #set variable blank to avoid endless loop
 newsfeed1=""
@@ -208,7 +218,8 @@ cat $workdir/optional2-news22.txt | sed 's/\%/\\%/g' >> $workdir/optional2-news.
 
 # Generate Video
 
-ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio5" -shortest -vf "drawtext=textfile='$workdir/optional2-news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $output/news-v3.mp4
+ffmpeg -y -f lavfi -i color=$newsbackground1:$videoresolution -stream_loop -1 -i "$audio5" -shortest -vf "drawtext=textfile='$workdir/optional2-news.txt': fontfile=/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf: x=(w-text_w)/2:y=h-$textspeed*t: fontcolor=$newstextcolour1: fontsize=W/40:"  -pix_fmt yuv420p -c:a copy -t $newsduration $wirkdir/news-v3.mp4
+ffmpeg -y -i $workdir/news-v3.mp4 -vf "fade=t=in:st=0:d=$newsvideofadeinduration,fade=t=out:st=$newsvideofadeoutstart:d=$newsvideofadeoutduration" -af "afade=t=in:st=0:d=$newsaudiofadeinduration,afade=t=out:st=$newsaudiofadeoutstart:d=$newsaudiofadeoutduration" $output/news-v3.mp4
 touch $output/news-v3.mp4
 #set variable blank to avoid endless loop
 newsfeed2=""