+++ helper-scripts/weather.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#V0.0.17 - Beta
+#V0.0.16 - Beta
 
 # load in configuration variables
 . config-temp.conf
@@ -14,12 +14,17 @@ processweather1=$(echo $processweather | tr '[:upper:]' '[:lower:]')
 if [[ $processweather1 = yes ]]
 then
 
-#weather
+  if [[ $audioamount = 1 ]]
+  then
+randomNumber=1
+randomNumber1=1
+randomNumber2=1
+else
+  randomNumber=$(shuf -i 1-$audioamount -n 1 --repeat)
+  randomNumber1=$(shuf -i 1-$audioamount -n 1 --repeat)
+  randomNumber2=$(shuf -i 1-$audioamount -n 1 --repeat)
+fi
 
-#audio
-randomNumber=$audionumber
-randomNumber1=$audionumber
-randomNumber2=$audionumber
 audio=$(head -n $randomNumber $workdir/music.txt | tail -n 1)
 audio1=$(head -n $randomNumber1 $workdir/music.txt | tail -n 1)
 audio2=$(head -n $randomNumber2 $workdir/music.txt | tail -n 1)
@@ -65,13 +70,19 @@ curl v2.wttr.in/${cityurl}.png$weathermeasurement --output $weatherdir/v2.png
 curl v3.wttr.in/${stateurl}.png$weathermeasurement --output $weatherdir/v3.png
 wait
 
+# Maths for fade
+weathervideofadeoutstart=$(echo `expr $videolength1 - $weathervideofadeoutduration` | bc)
+weatheraudiofadeoutstart=$(echo `expr $videolength1 - $weatheraudiofadeoutduration` | bc)
 #make video
 
-ffmpeg -y -f lavfi -i color=$background:$videoresolution -i $weatherdir/v1.png -stream_loop -1 -i "$audio" -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $output/weather-v1.mp4
+ffmpeg -y -f lavfi -i color=$background:$videoresolution -i $weatherdir/v1.png -stream_loop -1 -i "$audio" -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $workdir/weather-v1.mp4
+ffmpeg -y -i $workdir/weather-v1.mp4 -vf "fade=t=in:st=0:d=$weathervideofadeinduration,fade=t=out:st=$weathervideofadeoutstart:d=$weathervideofadeoutduration" -af "afade=t=in:st=0:d=$weatheraudiofadeinduration,afade=t=out:st=$weatheraudiofadeoutstart:d=$weatheraudiofadeoutduration" $output/weather-v1.mp4
 touch $output/weather-v1.mp4
-ffmpeg -y -f lavfi -i color=$background1:$videoresolution -i $weatherdir/v2.png -stream_loop -1 -i "$audio1" -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $output/weather-v2.mp4
+ffmpeg -y -f lavfi -i color=$background1:$videoresolution -i $weatherdir/v2.png -stream_loop -1 -i "$audio1" -shortest -filter_complex "[1]scale=iw*1:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $workdir/weather-v2.mp4
+ffmpeg -y -i $workdir/weather-v2.mp4 -vf "fade=t=in:st=0:d=$weathervideofadeinduration,fade=t=out:st=$weathervideofadeoutstart:d=$weathervideofadeoutduration" -af "afade=t=in:st=0:d=$weatheraudiofadeinduration,afade=t=out:st=$weatheraudiofadeoutstart:d=$weatheraudiofadeoutduration" $output/weather-v2.mp4
 touch $output/weather-v2.mp4
-ffmpeg -y -f lavfi -i color=$background2:$videoresolution -i $weatherdir/v3.png -stream_loop -1 -i "$audio2" -shortest -filter_complex "[1]scale=iw*0.9:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $output/weather-v3.mp4
+ffmpeg -y -f lavfi -i color=$background2:$videoresolution -i $weatherdir/v3.png -stream_loop -1 -i "$audio2" -shortest -filter_complex "[1]scale=iw*0.9:-1[wm];[0][wm]overlay=x=(W-w)/2:y=(H-h)/2" -pix_fmt yuv420p -c:a copy -t $videolength $workdir/weather-v3.mp4
+ffmpeg -y -i $workdir/weather-v3.mp4 -vf "fade=t=in:st=0:d=$weathervideofadeinduration,fade=t=out:st=$weathervideofadeoutstart:d=$weathervideofadeoutduration" -af "afade=t=in:st=0:d=$weatheraudiofadeinduration,afade=t=out:st=$weatheraudiofadeoutstart:d=$weatheraudiofadeoutduration" $output/weather-v3.mp4
 touch $output/weather-v3.mp4
 #end weather
 ./news.sh